This API is developed to answer the farming related queries. It gives JSON as an output. Developed in PHP.
# farming-QA-system--API
This API is developed to answer the farming
related queries. It gives JSON as an output. Developed in PHP.

The
format of this API is http://freakengineers.com/mtp/niks/index.php (base
URL).

The response is encoded in JSON.
Here are the possible
parameters placed in the request URL.
$query=N //Mandatory, N (string)
represents the query from the user.
$lcount=M //Optional, M represents
the no. of lines in summary.

Example-1
If the API request is made to
http://freakengineers.com/mtp/niks/index.php?query=iwanttoplantwatermelon
i.e. if the lcount parameter is not present. So for such cases we have
designed a smart summarization method.
Smart Summarisation - In this
first the length of the original text will be calculated, then according
to that, based on some selfmade rules, the length of the summary will be
decided by the system.

Example-2
If the API request is made to
http://freakengineers.com/mtp/niks/index.php?query=havetoplantladyfinger&lcount=3,5,7
Value
of lcount - lcount can take maximum upto three values, which means like
here, if we pass lcount=3,5,7 then the summarization will be shown in 3
lines, 5 lines and 7 lines. Also the smart summarization (default) will
be shown. The values of lcount should be separated by comma (,) and
there should not be any space.

Example-3
If the API request is made to
http://freakengineers.com/mtp/niks/index.php?query=growapple&lcount=2,3,7,5,1

i.e. If more then 3 values are passed in lcount, then it will show the
summary for initial 3 values along with the smart summarization
(default) and will display an error for the rest ones.
{"exception":"Max
line count exceeded"}

Output Link (JSON)-
nikunj.freakengineers.com/agro-api/index.php?query=how to plant
wheat&lcount=2,5